Fixing load function
Load function correctly reads data from file but it saves it in wrong order

+++ projekt_jipp/main.cpp
@@ -9,7 +9,7 @@
 int main()
 {
 	STACK_init(STUDENT_free, STUDENT_print);
-	size_t choice = 0;
+	int choice = 0;
 
 	while (choice >= INTREFACE_PUSH && choice <= INTERFACE_END)
 	{
